+++ tests/testthat/test-backend-.R
@@ -140,7 +140,11 @@ test_that("can translate nzchar", {
 # aggregates --------------------------------------------------------------
 
 test_that("all and any translated correctly", {
-  db <- memdb_frame(g = c(1, 1, 2, 2, 3, 3), x = c(0, 0, 0, 1, 1, 1))
+  db <- local_memdb_frame(
+    "df",
+    g = c(1, 1, 2, 2, 3, 3),
+    x = c(0, 0, 0, 1, 1, 1)
+  )
 
   sum_all_g <- db |>
     group_by(g) |>
@@ -193,31 +197,4 @@ test_that("default raw escapes translated correctly", {
 
   qry <- mf |> filter(x %in% !!L)
   expect_snapshot(qry)
-})
-
-# DDL ---------------------------------------------------------------------
-
-test_that("DDL operations generate expected SQL", {
-  con <- simulate_dbi()
-
-  expect_snapshot(sql_table_analyze(con, in_schema("schema", "tbl")))
-  expect_snapshot(sql_query_explain(con, sql("SELECT * FROM foo")))
-
-  expect_snapshot(sql_query_wrap(con, ident("table")))
-  expect_snapshot(sql_query_wrap(con, in_schema("schema", "tbl")))
-  expect_snapshot(sql_query_wrap(con, sql("SELECT * FROM foo")))
-
-  expect_snapshot(sql_table_index(con, in_schema("schema", "tbl"), c("a", "b")))
-  expect_snapshot(sql_table_index(
-    con,
-    in_schema("schema", "tbl"),
-    "c",
-    unique = TRUE
-  ))
-
-  expect_snapshot(sql_query_save(
-    con,
-    sql("SELECT * FROM foo"),
-    in_schema("temp", "tbl")
-  ))
 })