+++ tests/testthat/test-sql-build.R
@@ -1,5 +1,10 @@
 test_that("rendering table wraps in SELECT *", {
-  out <- copy_to_test("sqlite", tibble(x = 1), name = "test-sql-build")
+  out <- copy_to(
+    test_sqlite(),
+    tibble(x = 1),
+    name = "test-sql-build",
+    overwrite = TRUE
+  )
   expect_snapshot(out |> sql_render())
   expect_equal(out |> collect(), tibble(x = 1))
 })