+++ tests/testthat/test-tbl-lazy.R
@@ -23,17 +23,17 @@ test_that("has print method", {
 })
 
 test_that("support colwise variants", {
-  mf <- memdb_frame(x = 1:5, y = factor(letters[1:5]))
+  mf <- local_memdb_frame(x = 1:5, y = factor(letters[1:5]))
   exp <- mf |> collect() |> mutate(y = as.character(y))
 
   expect_message(
     mf1 <- dplyr::mutate_if(mf, is.factor, as.character),
     "on the first 100 rows"
   )
-  compare_tbl(mf1, exp)
+  expect_equal(collect(mf1), exp)
 
   mf2 <- dplyr::mutate_at(mf, "y", as.character)
-  compare_tbl(mf2, exp)
+  expect_equal(collect(mf2), exp)
 })
 
 test_that("base source of tbl_lazy is always 'df'", {