+++ tests/testthat/test-verb-arrange.R
@@ -1,17 +1,21 @@
 test_that("two arranges equivalent to one", {
-  skip("not clear whether test makes sense")
-  mf <- memdb_frame(x = c(2, 2, 1), y = c(1, -1, 1))
+  mf <- local_memdb_frame(x = c(2, 2, 1), y = c(1, -1, 1))
 
   mf1 <- mf |> arrange(x, y)
   mf2 <- mf |> arrange(y) |> arrange(x)
 
-  expect_warning(compare_tbl(mf1, mf2))
+  expect_equal(mf1 |> collect(), mf2 |> collect())
 })
 
 # sql_render --------------------------------------------------------------
 
 test_that("quoting for rendering ordered grouped table", {
-  db <- copy_to_test("sqlite", tibble(x = 1, y = 2), name = "test-verb-arrange")
+  db <- copy_to(
+    test_sqlite(),
+    tibble(x = 1, y = 2),
+    name = "test-verb-arrange",
+    overwrite = TRUE
+  )
   out <- db |> group_by(x) |> arrange(y) |> ungroup()
   expect_snapshot(sql_render(out))
   expect_equal(collect(out), tibble(x = 1, y = 2))
@@ -128,10 +132,8 @@ test_that("arranges captures DESC", {
 })
 
 test_that("multiple arranges combine", {
-  skip("not clear whether test makes sense")
   out <- lazy_frame(x = 1:3, y = 3:1) |> arrange(x) |> arrange(y)
-  out <- arrange(arrange(lazy_frame(x = 1:3, y = 3:1), x), y)
 
   sort <- lapply(op_sort(out), get_expr)
-  expect_equal(sort, list(quote(x), quote(y)))
+  expect_equal(sort, list(quote(y), quote(x)))
 })