+++ tests/testthat/test-verb-compute.R
@@ -1,41 +1,25 @@
-test_that("compute doesn't change representation", {
-  mf1 <- memdb_frame(x = 5:1, y = 1:5, z = "a")
-  compare_tbl(mf1, mf1 |> compute())
-  compare_tbl(mf1, mf1 |> compute() |> compute())
-
-  mf2 <- mf1 |> mutate(z = x + y)
-  compare_tbl(mf2, mf2 |> compute())
-})
-
 test_that("compute can create indexes", {
-  mfs <- test_frame(x = 5:1, y = 1:5, z = 10)
+  # integration test to ensure that indexes argument passed all the way through
+  db <- local_memdb_frame("db1", x = 5:1, y = 1:5, z = 10)
 
-  mfs |>
-    lapply(\(df) df |> compute(indexes = c("x", "y"))) |>
-    expect_equal_tbls()
+  db |> compute(indexes = c("x", "y"), name = "db2")
+  indices <- DBI::dbGetQuery(memdb(), "PRAGMA index_list('db2');")
+  expect_equal(indices$name, c("db2_y", "db2_x"))
+  expect_equal(indices$unique, c(0, 0))
 
-  mfs |>
-    lapply(\(df) df |> compute(indexes = list("x", "y", c("x", "y")))) |>
-    expect_equal_tbls()
-
-  mfs |>
-    lapply(\(df) df |> compute(indexes = "x", unique_indexes = "y")) |>
-    expect_equal_tbls()
-
-  mfs |>
-    lapply(\(df) {
-      df |> compute(unique_indexes = list(c("x", "z"), c("y", "z")))
-    }) |>
-    expect_equal_tbls()
+  db |> compute(unique_indexes = c("x", "y"), name = "db3")
+  indices <- DBI::dbGetQuery(memdb(), "PRAGMA index_list('db3');")
+  expect_equal(indices$name, c("db3_y", "db3_x"))
+  expect_equal(indices$unique, c(1, 1))
 })
 
 test_that("unique index fails if values are duplicated", {
-  mfs <- test_frame(x = 5:1, y = "a", ignore = "df")
-  lapply(mfs, function(.) expect_error(compute(., unique_indexes = "y")))
+  mf <- local_memdb_frame(x = 5:1, y = "a")
+  expect_error(compute(mf, unique_indexes = "y"))
 })
 
 test_that("index fails if columns are missing", {
-  mf <- memdb_frame(x = 1)
+  mf <- local_memdb_frame(x = 1)
   expect_snapshot({
     (expect_error(compute(mf, indexes = list(c("y", "x", "z"), "a"))))
     (expect_error(compute(mf, unique_indexes = list(c("y", "x", "z"), "a"))))
@@ -43,7 +27,7 @@ test_that("index fails if columns are missing", {
 })
 
 test_that("compute creates correct column names", {
-  out <- memdb_frame(x = 1) |>
+  out <- local_memdb_frame(x = 1) |>
     group_by(x) |>
     summarise(n = n()) |>
     compute() |>
@@ -53,7 +37,7 @@ test_that("compute creates correct column names", {
 })
 
 test_that("compute keeps window and groups", {
-  out <- memdb_frame(x = 1, y = 1) |>
+  out <- local_memdb_frame(x = 1, y = 1) |>
     window_order(x) |>
     group_by(x, y) |>
     summarise(n = n(), .groups = "drop_last") |>
@@ -66,7 +50,7 @@ test_that("compute keeps window and groups", {
 test_that("compute can handle named name", {
   con <- simulate_dbi()
   expect_equal(
-    memdb_frame(x = 1:10) |>
+    local_memdb_frame(x = 1:10) |>
       compute(name = c(x = unique_table_name())) |>
       collect(),
     tibble(x = 1:10)
@@ -74,8 +58,8 @@ test_that("compute can handle named name", {
 })
 
 test_that("compute can handle schema", {
-  df <- memdb_frame(x = 1:10)
-  withr::defer(DBI::dbRemoveTable(remote_con(df), "db1"))
+  df <- local_memdb_frame("db1", x = 1:10)
+  withr::defer(DBI::dbRemoveTable(memdb(), "db1"))
 
   expect_equal(
     df |>
@@ -93,13 +77,13 @@ test_that("compute can handle schema", {
 })
 
 test_that("compute(temporary = FALSE) without a name is deprecated", {
-  df <- memdb_frame(x = 1:10)
+  df <- local_memdb_frame(x = 1:10)
 
   expect_snapshot_warning(df |> compute(temporary = FALSE))
 })
 
 test_that("sorting preserved across compute", {
-  df1 <- memdb_frame(x = sample(10)) |> arrange(x)
+  df1 <- local_memdb_frame(x = sample(10)) |> arrange(x)
 
   df2 <- compute(df1)
   expect_equal(get_expr(op_sort(df2)[[1]]), quote(x))